+++ src/app/(main)/(contents)/posting-main/page.tsx
@@ -1,20 +1,15 @@
 'use client';
 import Button from '@/components/common/Button';
 import PostingFilter from '@/components/PostingPage/PostingMain/PostingFilter';
+import PostingList from '@/components/PostingPage/PostingMain/PostingList';
 import { useUserStore } from '@/store/userStore';
-import { createClient } from '@/supabase/client';
-import { Post, User } from '@/types';
-import Image from 'next/image';
-import Link from 'next/link';
 import { useRouter } from 'next/navigation';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 const PostingMainPage = () => {
-  const [posts, setPosts] = useState<Post[]>();
   // const [user, setUser] = useState<User[]>();
   const [selectedCategory, setSelectedCategory] = useState('전체 글 보기');
   // const { user, setUser } = useUserStore((state) => state);
-  const supabase = createClient();
   const router = useRouter();
   // const [currentPage, setCurrentPage] = useState(1);
   // const [postsPerPage] = useState(4);
@@ -25,25 +20,9 @@ const PostingMainPage = () => {
 
   // const paginate = (pageNumber: number) => setCurrentPage(pageNumber);
 
-  const getPosts = async () => {
-    const { data: posts, error } = await supabase.from('posts').select('*');
-    if (error) {
-      console.log('getPost error', error);
-      return;
-    }
-    setPosts(posts);
-  };
-
   const writePost = () => {
     router.push('/posting');
   };
-
-  useEffect(() => {
-    getPosts();
-  }, []);
-
-  const filteredPosts =
-    selectedCategory === '전체 글 보기' ? posts : posts?.filter((post) => post.category === selectedCategory);
 
   return (
     <main className="min-h-screen">
@@ -63,38 +42,7 @@ const PostingMainPage = () => {
         </div>
         <div className="border border-solid rounded-xl border-gray300 w-[1032px] ml-20 px-10 py-6">
           <h2 className="mb-4 text-2xl text-gray900 font-medium">건강한 식단 이야기</h2>
-          <ul>
-            {filteredPosts?.map((item, index: number) => (
-              <li
-                key={index}
-                className={`${
-                  index < filteredPosts.length - 1 ? 'border-b' : ''
-                }   border-gray200 pb-4 mb-4 cursor-pointer`}
-              >
-                {' '}
-                <Link href={`/posting-detail/${item?.id}`} className="flex">
-                  <Image
-                    src={item.image_url[0]}
-                    alt=""
-                    width={128}
-                    height={128}
-                    className="!w-[128px] !h-[128px] rounded-lg"
-                  />
-                  <div className="flex flex-col justify-between ml-5">
-                    <div>
-                      <span className="text-sm font-semibold text-primary600 mb-2">{item.category}</span>
-                      <p className="text-gray900 font-semibold">{item.title}</p>
-                      <p className="line-clamp-2">{item.content}</p>
-                    </div>
-                    {/* <div className="flex justify-between">
-                    <p>{item?.user_id === user?.userId ? user?.n}</p>
-                    <p>{item.date}</p>
-                  </div> */}
-                  </div>
-                </Link>
-              </li>
-            ))}
-          </ul>
+          <PostingList selectedCategory={selectedCategory} setSelectedCategory={setSelectedCategory} />
         </div>
         {/* <div className="flex justify-center mt-4">
           <button className={`px-3 py-1`} onClick={() => paginate(currentPage - 1)}>
